fix: security hardening round 2 — applicationId, API key timing, privacy, rate limiter
Round 2 security fixes from second codebase review:

Server:
- hmac.compare_digest() for constant-time API key comparison (timing attack fix)
- Health check returns generic 'unavailable' instead of leaking DB error details
- Rate limiter evicts stale IPs every 5 minutes (memory leak fix)
- PIN 4-digit limitation documented as known trade-off

App:
- Fix applicationId mismatch: Kotlin sources moved to com.mekenet.app
- All platform configs updated (Android, iOS, macOS, Linux)
- AndroidManifest receiver uses fully-qualified class name
- failed_parses stores SHA-256 hash instead of raw SMS text (privacy fix)
- SharedPreferences cached in CategoryService and SyncService

Docs:
- ARCHITECTURE.md updated with all round 2 fixes
- New test counts: 40 Flutter + 32 Server = 72 total

+++ server/models/transaction.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Float, DateTime, Boolean
+from sqlalchemy import Column, String, Numeric, DateTime
 from sqlalchemy.sql import func
 from db.database import Base
 
@@ -8,7 +8,7 @@ class Transaction(Base):
 
     id = Column(String, primary_key=True)
     type = Column(String, nullable=False)  # "income" or "expense"
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     description = Column(String, nullable=False)
     date = Column(String, nullable=False)  # YYYY-MM-DD
     category = Column(String, default="uncategorized")
